+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [1.38.7] - 2026-02-14
+
+### Added
+- legacy code purge + test fix
+
+## [1.38.6] - 2026-02-14
+
+### Added
+- legacy code purge + test fix
+
 ## [1.38.5] - 2026-02-14
 
 ### Bug Fixes